requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.9
Pygments==2.17.0
rich==13.7.0
toml==0.10.2
